// frontend/shared/types/api.ts
export interface CompanyUpdatePayload {
    name?: string;
    street_address?: string;
    city?: string;
    state?: string;
    zip_code?: string;
    country?: string;
    phone_number?: string;
}